Show timestamps of comments
When a new comment or reply is added, the current timestamp is stored (with
TS_MW format) and then rendered upon reviewing. At the moment when timestamps
are available they are shown right after the author name on the same line with
a space in between. To show older comments for which no timestamp was saved
some backwards-compatibility support was added.

The timestamp is *not* shown by the JavaScript when a comment or reply is first
made, only after the page reloads - an existing TODO note to make the two
displays match regarding the user was updated to note that timestamps should
be shown too.

KLAR-39