Make sure visual editor reloads page after edit so that annotations show up
VisualEditor will normally just alter the DOM based on api output. Instead
we make force a reload in order to ensure annotations show up properly.
This is a bit of a hacky solution, but it is hard to adjust the DOM
to ensure everything is correct after edit as visual editor does not
remove old annotations when inserting the new API html.

Change-Id: I42f2d1cfb3c126f1be5af195d158a8738df95816