Fix exception when highlight spans multiple child elements
There was a case where it would try and insert a text element
earlier in the tree. Remex doesn't allow inserting text elements
before another element if it is not the last child of the parent
element. As a work around, we insert the element elsewhere in the
doc until we are ready for it, then reparent it to the correct place.

Change-Id: I708dc04d1a37dd3d6eb75b0873f5f892ab1c2da7